refactor(server): shared listener package so CardDAV is bounded too
Review caught that the "every listener" claim was one listener short. CardDAV built its own http.Server with read and write timeouts only, and its handler chain had no cross-origin guard, while the docs and the constructor's own comment asserted listener-wide coverage. The bounds and the guard now live in a package of their own, internal/server/listen, that both api and carddav import, and CardDAV's bind and handler chain go through them. The guard sits after Basic auth in CardDAV because the forgery it stops is a browser replaying cached credentials on a cross-site request: the credentials pass, the origin does not.

Two wording fixes ride along. ReadTimeout covers the whole request read, headers included, so the header timeout's rationale now says what it actually adds: a tighter clock for the phase that has no legitimate reason to be slow. And the compat spec called the shim keys per-client; they are one key per surface with no per-client revocation, and the spec now says so in both places.
