+++ internal/server/api/ollama_server.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/nugget/thane-ai-agent/internal/platform/logging"
 	"github.com/nugget/thane-ai-agent/internal/runtime/agent"
+	"github.com/nugget/thane-ai-agent/internal/server/listen"
 )
 
 // OllamaServer is a dedicated server for Ollama-compatible API endpoints.
@@ -81,9 +82,9 @@ func (s *OllamaServer) Start(ctx context.Context) error {
 
 	// Auth and the cross-origin guard sit inside logging so rejected
 	// requests still produce access-log lines with their 401/403 status.
-	s.server = newHTTPServer(
+	s.server = listen.NewServer(
 		fmt.Sprintf("%s:%d", s.address, s.port),
-		s.withLogging(rejectCrossOriginWrites(s.logger, ollamaAuth(s.apiKey, mux))),
+		s.withLogging(listen.RejectCrossOriginWrites(s.logger, ollamaAuth(s.apiKey, mux))),
 		30*time.Second,
 		300*time.Second, // Long for slow models
 	)
