+++ internal/server/api/server.go
@@ -30,6 +30,7 @@ import (
 	"github.com/nugget/thane-ai-agent/internal/runtime/archivist"
 	looppkg "github.com/nugget/thane-ai-agent/internal/runtime/loop"
 	"github.com/nugget/thane-ai-agent/internal/server/legacyroute"
+	"github.com/nugget/thane-ai-agent/internal/server/listen"
 	"github.com/nugget/thane-ai-agent/internal/server/openapi"
 	"github.com/nugget/thane-ai-agent/internal/state/contacts"
 	"github.com/nugget/thane-ai-agent/internal/state/memory"
@@ -582,9 +583,9 @@ func (s *Server) Start(ctx context.Context) error {
 	// when ollama_api.enabled is true in config. Use RegisterOllamaRoutes()
 	// only if you need single-port operation.
 
-	s.server = newHTTPServer(
+	s.server = listen.NewServer(
 		fmt.Sprintf("%s:%d", s.address, s.port),
-		s.withLogging(rejectCrossOriginWrites(s.logger, limitRequestBody(nativeMaxBodyBytes, mux))),
+		s.withLogging(listen.RejectCrossOriginWrites(s.logger, listen.LimitRequestBody(nativeMaxBodyBytes, mux))),
 		30*time.Second,
 		120*time.Second, // Long for streaming responses
 	)
