+++ scripts/architecture.baseline
@@ -1,4 +1,4 @@
-packages=70
+packages=71
 interfaces=88
 large_files=59
 set_mutators=124
